+++ examples/pitch_plots/plot_hexbin.py
@@ -41,8 +41,8 @@ hexmap = pitch.hexbin(df_false9.x, df_false9.y, ax=ax, edgecolors='#f4f4f4',
 
 ##############################################################################
 # Load a custom font.
-URL = 'https://raw.githubusercontent.com/google/fonts/main/apache/roboto/Roboto%5Bwdth,wght%5D.ttf'
-URL2 = 'https://raw.githubusercontent.com/google/fonts/main/apache/robotoslab/RobotoSlab%5Bwght%5D.ttf'
+URL = 'https://raw.githubusercontent.com/google/fonts/main/apache/robotomono/RobotoMono[wght].ttf'
+URL2 = 'https://raw.githubusercontent.com/google/fonts/main/apache/robotoslab/RobotoSlab[wght].ttf'
 robotto_regular = FontManager(URL)
 robboto_bold = FontManager(URL2)